visual-biopsy/linux_brain/requirements.txt
pygame>=2.5.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyserial>=3.5
